+++ TwitchAPI/old/main.py
@@ -1,0 +1,5 @@
+from bot import Bot
+
+if __name__ == "__main__":
+    bot = Bot()
+    bot.run()
